Rm unnecessary submodule
